Feat: Implement reviews screen and update review model
This commit introduces a new screen to display user reviews and updates the `ReviewModel` and `Review` entity.

- Added `ReviewsScreen` to display a list of reviews using `ReviewCard` widgets.
- Created `ReviewCard` widget to display individual review details including reviewer name, rating, review text, and date.
- Modified `ReviewModel` to parse `rating` as `int` instead of `double`.
- Updated `Review` entity to expect `rating` as `int`.

+++ lib/features/review/domain/entities/review.dart
@@ -3,7 +3,7 @@ class Review {
   final DateTime updatedAt;
   final int id;
   final int courseId;
-  final double rating;
+  final int rating;
   final String comment;
 
   Review({
